backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
pydantic
openai
python-dotenv
